requirements.txt
annotated-types==0.6.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
en-core-web-lg==3.7.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
langcodes==3.3.0
lxml==5.2.1
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==23.3.1
preshed==3.0.9
py-readability-metrics==1.4.5
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
readability-lxml==0.8.1
-e ./readability2  # Editable install from local directory
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
trimesh==4.4.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.2
wheel==0.41.2
